Relocant

Description

A PostgreSQL migration library

Documentation

data Script

A migration script. Most users would get them by running readScripts, but those wanting a more fine-grained control would use readScript.

Constructors

Script

Fields

id :: ID

if XXX-YYYY.sql is the whole filename, then XXX is the id; see parseFilePath

name :: Name

if XXX-YYYY.sql is the whole filename, then XXX-YYYY is the name

content :: Content

readScripts :: FilePath -> IO [Script]

Read migration Scripts from a directory. It only tries to read paths ending with the .sql extension and doesn't recurse into subdirectories.

Note: the directory must exist

readScript :: FilePath -> IO Script

Read migration Scripts from a file. Useful when you need a more fine-grained control over which paths are read then what readScripts provides.

data ConnectionString

PostgreSQL connection string. (This is a newtype over ByteString)

for syntax, see: https://hackage.haskell.org/package/postgresql-simple/docs/Database-PostgreSQL-Simple.html#v:connectPostgreSQL

connect :: ConnectionString -> Table -> IO Connection

Connect to the DB using the given ConnectionString and initialize the migrations table.

Note: If you get your Connection from elsewhere, remember to call init yourself.

withLock :: Table -> Connection -> IO a -> IO a

Use pg_advisory_{lock,unlock} to restrict access to the migrations table. The given table is used to generate the lock's ID, so that in the unlikely case where you have multiple migrations tables in your database you can lock them separately.

withTryLock :: Table -> Connection -> (Bool -> IO a) -> IO a

A non-blocking version of withLock. True is passed to the callback if the lock was successfully acquired, False otherwise.

data Applied

An applied migration. Generally, it's created be either running a migration Script with apply or getting the records from the DB with getApplied or getAppliedByID.

Constructors

Applied

Fields

id :: ID

name :: Name

content :: Content

appliedAt :: At

durationS :: Duration

how long it took to apply the migration script

data ID

Migration ID.

data Name

Migration Name.

data Content

Migration content and its checksum.

data Duration

How long something took, in seconds. Use mempty for a 0 seconds Duration.

data Table

relocant's migration table name.

defaultTable :: Table

The default table name, which is "public.relocant_migration".

getApplied :: Table -> Connection -> IO [Applied]

Retrieve all Applied migrations' records from the DB.

getAppliedByID :: ID -> Table -> Connection -> IO (Maybe Applied)

Retrieve the specific Applied migration's record from the DB.

merge :: [Applied] -> [Script] -> Merged

Merge scripts and applied migrations, trying to discover inconsistencies and/or migration scripts to apply.

mergeAll :: FilePath -> Table -> Connection -> IO Merged

A convenience function that gets all applied migrations' records from the DB and merges them together with the migration scripts from a given directory.

data Merged

The result of merging Scripts and Applied migrations.

Constructors

Merged

Fields

unrecorded :: [Script]

A script that does not have a corresponding recorded migration, when there is a recorded migration with a higher ID

scriptMissing :: [Applied]

A recorded migration that does not have a corresponding script

contentMismatch :: [ContentMismatch]

A recorded migration and a script have the same ID but different content

unapplied :: [Script]

An unapplied script that has a higher ID than any recorded migration

data ContentMismatch

The applied migration and its purported script differ in content.

Constructors

ContentMismatch

Fields

expected :: Applied

butGot :: Script

canApply :: Merged -> Bool

No problems have been discovered after the merge.

converged :: Merged -> Bool

No problems have been discovered after the merge, and there are no migration scripts to apply.

apply :: Script -> Connection -> IO Applied

Run a Script against the database, get an Applied migration back if successful.

Note: You probably want to run this together with record in a single transaction.

record :: Applied -> Table -> Connection -> IO ()

Record a successfully Applied migration.

Note: You probably want to run this together with apply in a single transaction.

applyAll :: Merged -> Table -> Connection -> IO ()

A convenience function that applies all unapplied migrations' scripts and records their application in the DB. Each script is run in a separate transaction.

Re-exports (postgresql-simple)

data Connection

withTransaction :: Connection -> IO a -> IO a

Execute an action inside a SQL transaction.

This function initiates a transaction with a "begin transaction" statement, then executes the supplied action. If the action succeeds, the transaction will be completed with commit before this function returns.

If the action throws any kind of exception (not just a PostgreSQL-related exception), the transaction will be rolled back using rollback, then the exception will be rethrown.

For nesting transactions, see withSavepoint.