Relocant.Applied

Description

This module deals with applied migrations and their records in the DB.

Documentation

data Applied

An applied migration. Generally, it's created be either running a migration Script with apply or getting the records from the DB with getApplied or getAppliedByID.

Constructors

Applied

Fields

id :: ID

name :: Name

content :: Content

appliedAt :: At

durationS :: Duration

how long it took to apply the migration script

getApplied :: Table -> Connection -> IO [Applied]

Retrieve all Applied migrations' records from the DB.

getAppliedByID :: ID -> Table -> Connection -> IO (Maybe Applied)

Retrieve the specific Applied migration's record from the DB.

record :: Applied -> Table -> Connection -> IO ()

Record a successfully Applied migration.

deleteAll :: Table -> Connection -> IO ()

Delete all Applied migrations' records from the DB.

deleteByID :: ID -> Table -> Connection -> IO Bool

Delete the specific Applied migration's record from the DB.