Relocant.DB

Description

This module manages PostgreSQL connections and relocant's DB schema.

Documentation

data ConnectionString

PostgreSQL connection string. (This is a newtype over ByteString)

for syntax, see: https://hackage.haskell.org/package/postgresql-simple/docs/Database-PostgreSQL-Simple.html#v:connectPostgreSQL

data Table

relocant's migration table name.

defaultTable :: Table

The default table name, which is "public.relocant_migration".

connect :: ConnectionString -> Table -> IO Connection

Connect to the DB using the given ConnectionString and initialize the migrations table.

init :: Connection -> Table -> IO ()

Initialize the migrations table.

withLock :: Table -> Connection -> IO a -> IO a

Use pg_advisory_{lock,unlock} to restrict access to the migrations table. The given table is used to generate the lock's ID, so that in the unlikely case where you have multiple migrations tables in your database you can lock them separately.

withTryLock :: Table -> Connection -> (Bool -> IO a) -> IO a

A non-blocking version of withLock. True is passed to the callback if the lock was successfully acquired, False otherwise.

lock :: Table -> Connection -> IO ()

Use pg_advisory_lock to lock the database, restricting access to the migrations table. The given table is used to generate the lock's ID, so that in the unlikely case where you have multiple migrations tables in your database you can lock them separately.

tryLock :: Table -> Connection -> IO Bool

A non-blocking version of lock. Returns True if the lock was successfully acquired, False otherwise.

unlock :: Table -> Connection -> IO Bool

Use pg_advisory_unlock to unlock the database.

dumpSchema :: IO ()