Relocant.Merge

Description

This modules deals with merging scripts and applied migrations, and finding inconsistencies.

Documentation

data Merged

The result of merging Scripts and Applied migrations.

Constructors

Merged

Fields

unrecorded :: [Script]

A script that does not have a corresponding recorded migration, when there is a recorded migration with a higher ID

scriptMissing :: [Applied]

A recorded migration that does not have a corresponding script

contentMismatch :: [ContentMismatch]

A recorded migration and a script have the same ID but different content

unapplied :: [Script]

An unapplied script that has a higher ID than any recorded migration

data ContentMismatch

The applied migration and its purported script differ in content.

Constructors

ContentMismatch

Fields

expected :: Applied

butGot :: Script

merge :: [Applied] -> [Script] -> Merged

canApply :: Merged -> Bool

No problems has been discovered after the merge.

converged :: Merged -> Bool

No problems has been discovered after the merge, and there are no migration scripts to apply.