Relocant.Migration.Merge

Documentation

data Merged

Constructors

Merged

Fields

unrecorded :: [Script]

a script that does not have a corresponding recorded migration, when there is a recorded migration with a higher ID

scriptMissing :: [Migration]

a recorded migration that does not have a corresponding script

contentMismatch :: [ContentMismatch]

a recorded migration and a script have the same ID but different content

unapplied :: [Script]

an unapplied script that has a higher ID than any recorded migration

data ContentMismatch

Constructors

ContentMismatch

Fields

expected :: Migration

butGot :: Script

merge :: [Migration] -> [Script] -> Merged

canApply :: Merged -> Bool

converged :: Merged -> Bool