Relocant.Script

Description

This module deals with migrations that haven't yet been applied.

Documentation

data Script

A migration script. Most users would get them by running readScripts, but those wanting a more fine-grained control would use readScript.

Constructors

Script

Fields

id :: ID

if XXX-YYYY.sql is the whole filename, then XXX is the id; see parseFilePath

name :: Name

if XXX-YYYY.sql is the whole filename, then XXX-YYYY is the name

content :: Content

readScripts :: FilePath -> IO [Script]

Read migration Scripts from a directory. It only tries to read paths ending with the .sql extension and doesn't recurse into subdirectories.

Note: the directory must exist

readScript :: FilePath -> IO Script

Read migration Scripts from a file. Useful when you need a more fine-grained control over which paths are read then what readScripts provides.

apply :: Script -> Connection -> IO Applied

Run a Script against the database, get an Applied migration back if successful.

markApplied :: Script -> IO Applied

Get an Applied migration from a Script, without running it. This should not be normally used, but can be useful for fixing checksums after cosmetics updates of migration Scripts (such as adding comments, for example).

Extra

parseFilePath :: FilePath -> (ID, Name)

Get XXX as the ID and XXX-YYYY as the Name from XXX-YYYY.sql

Basically, the longest alphanumeric prefix of the basename is the ID, and the basename is the Name.

readContent :: FilePath -> IO Content

Get migration Script's content and its checksum.